+++ src/guestbooks/domain/useCases/GetGuestbookResponsesByGuestbookId.ts
@@ -1,19 +1,19 @@
 import { UseCase } from '../../../core/domain/useCases/UseCase'
-import { GuestbookResponse } from '../models/GuestbookResponse'
+import { GuestbookResponseSubset } from '../models/GuestbookResponse'
 import { IGuestbooksRepository } from '../repositories/IGuestbooksRepository'
 
-export class GetGuestbookResponsesByGuestbookId implements UseCase<GuestbookResponse[]> {
+export class GetGuestbookResponsesByGuestbookId implements UseCase<GuestbookResponseSubset> {
   constructor(private readonly guestbooksRepository: IGuestbooksRepository) {}
 
   /**
-   * Returns guestbook responses for one guestbook.
+   * Returns paginated guestbook responses for one guestbook.
    *
    * @param {number} guestbookId - Guestbook identifier.
    * @param {number} limit - Maximum number of responses to return.
    * @param {number} offset - Number of responses to skip.
-   * @returns {Promise<GuestbookResponse[]>}
+   * @returns {Promise<GuestbookResponseSubset>}
    */
-  async execute(guestbookId: number, limit = 10, offset = 0): Promise<GuestbookResponse[]> {
+  async execute(guestbookId: number, limit = 10, offset = 0): Promise<GuestbookResponseSubset> {
     return await this.guestbooksRepository.getGuestbookResponsesByGuestbookId(
       guestbookId,
       limit,